MWPW-151240: Moved caasUrl logic to utils.js
