+++ test/blocks/partner-news/partner-news.test.js
@@ -41,7 +41,7 @@ describe('partner-news block', () => {
     expect(partnerNewsWrapper.shadowRoot).to.exist;
     const partnerCardsCollection = partnerNewsWrapper.shadowRoot.querySelector('.partner-cards-collection');
     expect(partnerCardsCollection).to.exist;
-    expect(partnerCardsCollection.innerHTML).to.include('news-card');
+    expect(partnerCardsCollection.innerHTML).to.include('single-partner-card');
     const firstCard = partnerCardsCollection.querySelector('.card-wrapper');
     expect(firstCard.shadowRoot).to.exist;
     const searchBarWrapper = partnerNewsWrapper.shadowRoot.querySelector('.partner-cards-sidebar .search-wrapper');
